Conexion de arduino y base de datos exitosa

+++ src/main/java/co/edu/eafit/solarcardata/Ambiental.java
@@ -10,12 +10,12 @@ package co.edu.eafit.solarcardata;
  * @author Daniel
  */
 public class Ambiental extends Dato{
-    public float temperatura;
+    public double temperatura;
     public float radiacion;
     public float magViento;
     public String dirViento;
     
-    public Ambiental(float temperatura, float radiacion, float magViento, String dirViento){
+    public Ambiental(double temperatura, float radiacion, float magViento, String dirViento){
         this.temperatura=temperatura;
         this.radiacion=radiacion;
         this.magViento=magViento;
